Random enemy spawn
To use the random enemy spawn make one prefab for the position of the enemies. Do that for as many different enemy spawns you want.
Make a prefab out of the prefabs of the spawns you just made. In the Game Manager script line 12, change the maximum number of the random range of the prefChild to the number of spawners you made.