Checkpoints now work, game resets properly, bug fix, check walls level 1
On level one I experimented with changing the layer which the walls are drawn so you can go behind them. Changed the "level 3" and "level 5" to "Checkpoint 1" and "Checkpoint 2'. (This naming is crucial for the code) Coded in MainMenu.cs and Retry.cs to load the correct room after dying or clicking the play button. Added to the "End" and the "TitleScreen" scenes an object called Player with the player script attached and the GameManager prefab so we could save/reset/load. If you kill the boss, you go to the title screen and the game Hard Resets the stats. Some extra coding to make everything come together. Fixed a bug that made the player have less than 0 corruption.