Fire Weapon Primary attack
Clicking and holding left click starts charging a fireball making it bigger and stronger. Releasing left click shoots the fireball. There is a limit to how much you can charge the fire ball and if you reach it, it automatically shoots the fireball. Currently I made the max fireball damage to be 40 because 40 is the hp of the Draugr. The damage range can be changed from the bullet component on the BulletprefabFIRE prefab.