Fire Weapon Primary attack 1.01
Connected the cooldown bar to the Fire Weapon script to all scenes. Left a Debug.Log on the bullet script so we see what was the amount of damage the Fireball had when released. Bugs/Features: The player doesn't stop when charging the fireball and the fireball doesn't follow the player either, so for now if you don't stop moving while charging the fireball will stay at the position the player was when he first used left click. The fireball does not rotate towards where the mouse is after it is already instantiated.