Final Game Hierarchy, Corruption, Mist
All the final level scenes were moved under "_The final game/scenes". Folders were created for the prefabs for the enemies that will spawn on each level and the mist that will also spawn on each level. It was organized that way so that it is easier to navigate and to bunch everything for the final level making together in one place on the top of all the project files.

The corruption mechanic was completely removed from the game (hopefully with 100% success)

The mist now persists throughout the level.